Implemented the Proof of use
At first, it was some trollish lines of code: each transaction would generate 1 token. Then, I thought about the consequences of such thing...

It means there is native liquidity between Catether and Ethereum. Let's say you decide to transfer 0.00000001 CATE to someone: you generate (almost) 1 CATE. However, this transaction costs gas (around 20 000 gas). Since this token is generated by the EVM, it means EVM is "asking" for 20000-or-so gas to mint 1 cat.
And every time you bid a price in ether for your gas (your Gwei/gas is actually a sort of "bid" order on the TX-mining market of miners), each time someone make a transaction in CATE, it "matches" an "order".
That's a lot of quotemarks, but it's quite serious. It creates native liquidity.

Furthermore: let's say you decide to not make a donation (you selfish bastard), and instead mint it for yourself. Well, you are reimbursing part of your transaction in CATE, so, it's basically always cheaper to make transaction in Catether.
The idea is that, once there's enough usage, we will have a...

Free.

medium of transaction for the Ethereum blockchain, pegged to the price of the transaction in gas.

Canged the numbers, since it's an 8-decimal coin, wee need to add  10'000'000 atomic units to credit one CATE.
I approximated that there could be, at max, a billion transaction in a year. In its later stage, the mining process will generate around 5.6B CATES/year.
To that emmission rate, we should now add "X"*1000'000'000 as an upper bound to the emmission rate.

I choosed the Golden Number(1.61803400).

It means that, if we hit 1 Billion market cap at some point, the CATE transfer will be free.

Period.
Hope you enjoyed the little explanation.