Actually, it's quite costly costly to mint blocks.
Having a short target time for blocks only improves one thing : mining variance.
The more blocks are slow, the less gas it costs in minting per hour. Also, with a 1 minute block-time, there's a lot of tx collisions between miners and pools. So, we need more blocks, so that the minting process is spread on more blocks, thus decreasing minting collisions.

I've set it to 3.141 because this number's a cutie Pie (ahah), and also because, in average, it means you will have to wait for Ceiling(BlockNumber/Blocktime), with BlocNumber/Blocktime = 3.141, so, in fine it means that, 85.9% we're on a 16-block average, and 14.1% of the time, 15 blocks.

Anyway. This little fix reduces minting costs for the network, and also reduce the number of failed transactions.