Changed the initial difficulty.
It seemed like the difficulty was so low for GPU miners that it caused overflows in the miner software. Since I wanted a flawless experience for my early miners, I made it so the launch can be done without any issue on the current mining softs.

Launching in a timely manner !