Much change, how commits, wow
Had to change a couple things.

First, the difficulty retarget: it was completely broken. I had to completely rewrite it, and now we're getting somewhere : the retarget difficulty can move up 20-fold in one hour, or 50 CATE blocks (which now have a timeTarget of 5 minute). So, a bit less than a 4x speed.
It mainly comes from that Ethereum network lag between a transfer broadcast and its confirmation, we will have to wait for Ethereum to scale.

In the meantime, it acts as an additional tampering on the first Tampering we do on the...

New EMA-retarget that has just been implemented.

Also, I lowered the number of decimals: 4 is fine, it allows for 1/10'000th increments, which should be more-than enough for a token that will cost no more than a few cents (unless transaction costs goes up like crazy)

To match that change, and also after a few tinkering, I changed the Donation reward: now, it will mint 1.0000 CATE, and will slice it in two (savage). First half will go for msg.sender, second half will go to a dynamically assigned address, always choosen by msg.sender.

That way, once an address receives CATE, you cannot reset the balance to 0 (helps tracking : just check for non-0 balances), and also, the ones actually using the Donation feature to actually donate their Cate to a Non-Governmental-Organization, or for research against Cancer ( ... I hope some people do this. It really is the actual purpose of all this, so that anyone can create money so that they can hapily use it for the Greater Good)... The ones donating to NGOs can now choose two levels of donation :

50 % donation: just change your Donation Address to your liking and let the magic do the rest

100% donation : let's say you really want to donate everything to an NGO: they just have to create a smart contract that implements the "TransferFrom" function, and set the contract's donation address to itself. Now they can make your transfer in your name and get 100% of the reward.
(additionnal gas costs may apply but hey, you're making something really close to micro-donations, natively, and that is awesome. Congratulation, kind-hearted lad. You did something good.)

Changed the minting reward halving for something a bit less "halvy" at first.
The main reasoning behind this is: broader distribution. If you just halve every X block, people starting to mine after a halving are missing out on literally half the supply. That's not great. So, for Catether it wil more be like "Tiering", with the first 9 months being close to 33% decrease between each eras, then three halving, then two additionnal tierings. It looks like a Gauss bell, in terms of Reward/Time value. Just like what usage scaling should look like.

Also, solution of the first block is 42.
ahah yes